+++ game/src/main/java/com/bship/games/endpoints/BoardsController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
-import java.math.BigInteger;
 import java.util.function.Function;
 import java.util.stream.Stream;
 
@@ -41,7 +40,7 @@ public class BoardsController implements BadRequestHandler {
             value = "/boards/{boardId}",
             produces = APPLICATION_JSON_UTF8_VALUE)
     @ResponseStatus(OK)
-    public Board placeShip(@PathVariable BigInteger boardId,
+    public Board placeShip(@PathVariable long boardId,
                            @Valid @RequestBody Ship ship) throws BoardValidation {
         return service.placeShip(boardId, ship);
     }
